backend/pyproject.toml
[tool.poetry]
name = "bookwise-backend"
version = "0.1.0"
description = "BookWise backend service"
authors = ["Your Name"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
sqlalchemy = "^2.0"
httpx = "^0.27.0"
openai = "^1.0"
reportlab = "^4.0"
python-dotenv = "^1.0"
slowapi = "^0.1"
structlog = "^24.0"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
black = "^24.0"
ruff = "^0.4"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
